pr comments + clean up
